+++ app/_components/browser-run-detail.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Client } from "eve/client";
+import { Client, type MessageStreamEvent } from "eve/client";
 import { ArrowLeftIcon, RefreshCwIcon } from "lucide-react";
 import { useRouter } from "next/navigation";
 import { useEffect, useRef, useState } from "react";
@@ -13,8 +13,19 @@ import {
 import { useBrowserRunGroups } from "@/app/_components/use-browser-run-groups";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
-import type { BrowserRunGroup, BrowserRunTask } from "@/lib/browser-run-store";
-import { runPersistedTask } from "@/lib/browser-task-runner";
+import {
+  measureBrowserTask,
+  readTaskCompletion,
+  terminalBrowserMessage,
+} from "@/lib/browser-benchmark";
+import {
+  type BrowserRunGroup,
+  type BrowserRunTask,
+  type BrowserRunTaskUpdate,
+  updateBrowserRunTask,
+} from "@/lib/browser-run-store";
+
+const taskTimeoutMs = 15 * 60_000;
 
 export function BrowserRunDetail({ groupId }: { readonly groupId: string }) {
   const router = useRouter();
@@ -179,6 +190,116 @@ async function runGroup(
   );
 }
 
+async function runPersistedTask(
+  client: Client,
+  groupId: string,
+  task: BrowserRunTask
+) {
+  const requestStartedAt = task.startedAt ?? Date.now();
+  const events: MessageStreamEvent[] = [];
+  let timeout: number | undefined;
+
+  const update = (taskUpdate: BrowserRunTaskUpdate) =>
+    updateBrowserRunTask(groupId, task.id, taskUpdate);
+
+  try {
+    if (task.status === "running" && task.sessionId) {
+      const session = client.sessions.attach(task.sessionId, {
+        streamIndex: 0,
+      });
+      for await (const event of session.stream({ startIndex: 0 })) {
+        events.push(event);
+        projectTaskEvents(events, requestStartedAt, update);
+      }
+    } else {
+      const { response } = await client.sessions.create({
+        message: task.prompt,
+      });
+      update({
+        sessionId: response.sessionId,
+        startedAt: requestStartedAt,
+        status: "running",
+      });
+      timeout = window.setTimeout(() => {
+        update({ terminalMessage: "Timed out after 15 minutes; cancelling…" });
+        void response.cancel();
+      }, taskTimeoutMs);
+
+      for await (const event of response) {
+        events.push(event);
+        projectTaskEvents(events, requestStartedAt, update);
+      }
+    }
+
+    completePersistedTask(events, requestStartedAt, update);
+  } catch (error) {
+    const completion = readTaskCompletion(events);
+    const metrics = measureBrowserTask(events, Date.now() - requestStartedAt);
+    update({
+      completedAt: Date.now(),
+      costComplete: metrics.costComplete,
+      costUsd: metrics.costUsd,
+      durationMs: metrics.durationMs,
+      status: completion?.status ?? "failure",
+      terminalMessage: completion?.message ?? toErrorMessage(error),
+    });
+  } finally {
+    if (timeout !== undefined) window.clearTimeout(timeout);
+  }
+}
+
+function projectTaskEvents(
+  events: readonly MessageStreamEvent[],
+  requestStartedAt: number,
+  update: (taskUpdate: BrowserRunTaskUpdate) => void
+) {
+  const event = events.at(-1);
+  if (!event || !shouldProjectEvent(event)) return;
+
+  const metrics = measureBrowserTask(events, Date.now() - requestStartedAt);
+  const completion = readTaskCompletion(events);
+  update({
+    costComplete: metrics.costComplete,
+    costUsd: metrics.costUsd,
+    durationMs: metrics.durationMs,
+    terminalMessage: completion?.message,
+  });
+}
+
+function completePersistedTask(
+  events: readonly MessageStreamEvent[],
+  requestStartedAt: number,
+  update: (taskUpdate: BrowserRunTaskUpdate) => void
+) {
+  const metrics = measureBrowserTask(events, Date.now() - requestStartedAt);
+  const completion = readTaskCompletion(events);
+  const fallbackMessage = terminalBrowserMessage(undefined, events);
+  update({
+    completedAt: Date.now(),
+    costComplete: metrics.costComplete,
+    costUsd: metrics.costUsd,
+    durationMs: metrics.durationMs,
+    status: completion?.status ?? "failure",
+    terminalMessage:
+      completion?.message ??
+      (fallbackMessage === "No terminal message"
+        ? "Task ended without calling complete_task."
+        : fallbackMessage),
+  });
+}
+
+function shouldProjectEvent(event: MessageStreamEvent) {
+  return (
+    event.type === "message.received" ||
+    event.type === "actions.requested" ||
+    event.type === "action.result" ||
+    event.type === "step.completed" ||
+    event.type === "turn.failed" ||
+    event.type === "turn.cancelled" ||
+    event.type === "session.failed"
+  );
+}
+
 function groupWallTime(group: BrowserRunGroup, now: number) {
   const starts = group.tasks.flatMap((task) =>
     task.startedAt === undefined ? [] : [task.startedAt]
@@ -203,4 +324,8 @@ const groupTimestampFormatter = new Intl.DateTimeFormat(undefined, {
 
 function formatGroupTimestamp(timestamp: string) {
   return groupTimestampFormatter.format(new Date(timestamp));
+}
+
+function toErrorMessage(error: unknown) {
+  return error instanceof Error ? error.message : "Task failed.";
 }
