+++ app/_components/browser-run-table.tsx
@@ -114,9 +114,7 @@ export function BrowserRunTable({
                   <div className="flex items-start gap-2">
                     <span className="min-w-0 flex-1 text-muted-foreground">
                       {task.terminalMessage ??
-                        (task.status === "running"
-                          ? (task.activity ?? "Running…")
-                          : "—")}
+                        (task.status === "running" ? "Running…" : "—")}
                     </span>
                     {task.sessionId ? (
                       <Button
