--- lib/task-timeout.ts
@@ -1,8 +1,0 @@
-export const browserTaskTimeoutMs = 15 * 60_000;
-
-export function remainingTaskTimeoutMs(
-  startedAt: number,
-  now = Date.now()
-): number {
-  return Math.max(0, browserTaskTimeoutMs - Math.max(0, now - startedAt));
-}
